+++ src/Tests/ContainerService.fs
@@ -737,32 +737,35 @@ let tests =
 
             Expect.equal osSKU "AzureLinux" "Incorrect osSKU value"
         }
-        test "AKS cluster with system identity and ACR access" {
+        test "AKS cluster with kubelet identity and ACR access" {
+            let kubeletMsi = createUserAssignedIdentity "kubeletIdentity"
             let myAcr = containerRegistry { name "mycontainerregistry" }
             let myAcrResId = (myAcr :> IBuilder).ResourceId
 
-            let myAks = aks {
-                name "aks-cluster"
-                service_principal_use_msi
-            }
-
-            let myAksResId = (myAks :> IBuilder).ResourceId
-
             let acrPullRoleNameExpr =
                 ArmExpression.create (
-                    $"guid(concat(resourceGroup().id, '{myAksResId.Name.Value}', '{Roles.AcrPull.Id}'))"
+                    $"guid(concat(resourceGroup().id, '{kubeletMsi.ResourceId.Name.Value}', '{Roles.AcrPull.Id}'))"
                 )
 
             let acrPullRole = {
                 Name = acrPullRoleNameExpr.Eval() |> ResourceName
                 RoleDefinitionId = Roles.AcrPull
-                PrincipalId = myAks.SystemIdentity.PrincipalId
+                PrincipalId = kubeletMsi.PrincipalId
                 PrincipalType = PrincipalType.ServicePrincipal
                 Scope = AssignmentScope.SpecificResource myAcrResId
-                Dependencies = Set [ myAksResId; myAcrResId ]
+                Dependencies = Set [ kubeletMsi.ResourceId ]
+            }
+
+            let myAks = aks {
+                name "aks-cluster"
+                service_principal_use_msi
+                kubelet_identity kubeletMsi
+                depends_on myAcr
+                depends_on_expression acrPullRoleNameExpr
             }
 
             let template = arm {
+                add_resource kubeletMsi
                 add_resource myAcr
                 add_resource myAks
                 add_resource acrPullRole
@@ -771,17 +774,19 @@ let tests =
             let json = template.Template |> Writer.toJson
             let jobj = Newtonsoft.Json.Linq.JObject.Parse(json)
 
-            // Verify AKS has system identity
-            let identity =
-                jobj.SelectToken("resources[?(@.name=='aks-cluster')].identity.type") |> string
+            // Verify AKS has kubelet identity configured
+            let kubeletIdentityClientId =
+                jobj.SelectToken(
+                    "resources[?(@.name=='aks-cluster')].properties.identityProfile.kubeletIdentity.clientId"
+                )
 
-            Expect.equal identity "SystemAssigned" "AKS cluster should have a SystemAssigned identity for ACR access"
+            Expect.isNotNull kubeletIdentityClientId "AKS cluster should have kubelet identity configured"
 
             // Verify role assignment exists in template
             let roleAssignments =
                 jobj.SelectTokens("resources[?(@.type=='Microsoft.Authorization/roleAssignments')]")
                 |> Seq.toList
 
-            Expect.hasLength roleAssignments 1 "Should have one role assignment in template"
+            Expect.hasLength roleAssignments 1 "Should have one role assignment for AcrPull in template"
         }
     ]